+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.3.33] - 2019-02-17
+### Added
+- Config to disable heart color looping [#155]
 ### Fixed
 - Crash when connecting to servers in some cases [#157]
 - Heart container drops not respecting doMobLoot game rule [#156]